Create ShareAccess controller and routing
