+++ app/javascript/components/forms/DeleteSharedAccessForm.jsx
@@ -5,11 +5,13 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faTrashAlt } from '@fortawesome/free-regular-svg-icons';
 import { useParams } from 'react-router-dom';
 import useDeleteSharedAccess from '../../hooks/mutations/rooms/useDeleteSharedAccess';
+import useRoom from '../../hooks/queries/rooms/useRoom';
 
-export default function DeleteSharedAccessForm({userId}) {
+export default function DeleteSharedAccessForm({ userId }) {
   const { register, handleSubmit } = useForm();
   const { friendlyId } = useParams();
-  const { onSubmit } = useDeleteSharedAccess(friendlyId);
+  const { data: room } = useRoom(friendlyId);
+  const { onSubmit } = useDeleteSharedAccess(room.id);
 
   return (
     <Form onSubmit={handleSubmit(onSubmit)} className="float-end pe-2">
