+++ app/javascript/hooks/queries/rooms/useShareableUsers.jsx
@@ -1,8 +1,8 @@
 import { useQuery } from 'react-query';
 import axios from 'axios';
 
-export default function useShareableUsers(friendlyId) {
-  return useQuery('getShareableUsers', () => axios.get(`/api/v1/rooms/${friendlyId}/shareable_users.json`, {
+export default function useShareableUsers(roomId) {
+  return useQuery('getShareableUsers', () => axios.get(`/api/v1/shared_accesses/${roomId}/shareable_users.json`, {
     headers: {
       'Content-Type': 'application/json',
       Accept: 'application/json',
