+++ app/javascript/hooks/queries/rooms/useSharedUsers.jsx
@@ -1,8 +1,8 @@
 import { useQuery } from 'react-query';
 import axios from 'axios';
 
-export default function useSharedUsers(friendlyId) {
-  return useQuery('getSharedUsers', () => axios.get(`/api/v1/rooms/${friendlyId}/shared_users.json`, {
+export default function useSharedUsers(roomId) {
+  return useQuery('getSharedUsers', () => axios.get(`/api/v1/shared_accesses/${roomId}/shared_users.json`, {
     headers: {
       'Content-Type': 'application/json',
       Accept: 'application/json',
